requirements.txt
gradio
langchain
langchain_community
langchain_core
langchain-openai
pandas
pyprojroot
python-dotenv
PyYAML
SQLAlchemy
openai
chromadb
pymongo

